requirements.txt
Django==5.0.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-filter==23.5
django-cors-headers==4.3.1
pandas==2.2.0
openpyxl==3.1.2
python-dotenv==1.0.1
Pillow==10.2.0 